[wip] Fix Envers compatibility with ImmutableType
Envers expects (simple) audited types to implement the BasicType
interface. This commit changes The ImmutableType to implement BasicType
rather than Type.

The commit is work in progress: It only applies the change to
hibernate-types-55 as a template. If the change is OK, it will be
applied to other hibernate-types-* as well.

Resolves: https://github.com/vladmihalcea/hibernate-types/issues/463